refactor: Separando o schema da função validate e validando a compatibilidade dos emails no próprio schema
